Update after rebasing to refactored models.
