+++ mreg/tests/base.py
@@ -1,7 +1,6 @@
 """Base class for Mreg tests."""
 
+
 def clean_and_save(entity):
     entity.full_clean()
     entity.save()
-
-
