+++ mreg/tests/test_model_loc.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
-
-from mreg.models import Host, Loc
+from mreg.models.host import Host
+from mreg.models.resource_records import Loc
 
 from .base import clean_and_save
