+++ mreg/tests/test_model_sshfp.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
-
-from mreg.models import Host, Sshfp
+from mreg.models.host import Host
+from mreg.models.resource_records import Sshfp
 
 from .base import clean_and_save
