+++ mreg/tests/test_model_txt.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
-
-from mreg.models import Host, Txt
+from mreg.models.host import Host
+from mreg.models.resource_records import Txt
 
 from .base import clean_and_save
